Disable DelayedJobs for the Capistrano role
DelayedJobs was already disabled by default for the basic installation. However once the Capistrano role was executed DelayedJobs was activated.

This stored emails in the DelayedJob queue and were not being delieved.

With this commit we are disabling DelayedJobs by default after executing the Capistrano role, making the default configuration for emails easier.

Once we setup the whole DelayedJobs infraestructure[1] we can enable DelayedJobs again.

[1] https://github.com/consul/docs/issues/54