Disable RMV auto updates
There is a bug[1] in the latest release of RVM (1.29.5). When running the script a second time, RMV was updating to the latest buggy release making the script not complete successfully.

With this commit we are making sure the the installer maintains RMV's version 1.29.4 and does not auto update to 1.29.5 in subsequent runs of the script.

[1] https://github.com/consul/installer/pull/63