Simplify Nginx installation
We are getting an error from the Nginx Galaxy as the configuration is only partial complete when the nginx module is called with the include_role command.

With this commit we are installing Nginx manually which gives us more control over when to restart the nginx server, and so to avoid these errors.