Integrate LetsEncrypt with Travis
With this commit we are generating a self-signed certificate to be used later on in specs.

Note that an alternative implementation would be to use a Pebble[1] server, which is a LetsEncrypt server used for testing purposes. However documenation is sparse and it has been complicated to setup.

Using a self signed ssl certificate is not ideal but it's a start.

[1] https://github.com/letsencrypt/pebble