Add specs for LetsEncrypt
Note that we are testing varios things here, it's not as easy to follow as a rspec spec, so to summarize:

When a domain is configured, we want to make sure we can:
- Access the site through https.
- Get redirected to https when trying to access through http.

When the domain is not configured, we want to make sure we can:
- Access the site through http.
- No access to https.

Note that the above specs will be executed on every run of the installer, even if it is not run in Travis. The next spec would only be run in Travis.

When we are in Travis, we want to make sure we can:
- Access the site through https
- Get redirected to https when trying to access through http.