Use capistrano folder structure
This way activating capistrano is much easier. For instance, we can
always use `consul/current/` as the release path, so we don't have to
update nginx nor puma configuration files.

Note we're deleting the log directory in order to be able to convert it
to a link. It might be better to remove it from CONSUL's repository
instead, but since right now it's present there, we need a way to deal
with it.