Split tasks to setup CONSUL
This way we differentiate tasks which exclusively operate with the
CONSUL folder and tasks which configure some other services.

In order to do so, we need to move the part of `letsencrypt` dealing
with the rails application to the `rails` role.

The `ruby` task is the exception. I wasn't sure what to use here.
Ideally we would clone the repository first and then install ruby based
on the version in `.ruby-version`, which is inside the CONSUL folder, so
I'm keeping it inside the services to setup CONSUL.