Revert change in RVM galaxy code
This code was changed in commit 35f859e0 because I thought we had to
change it to avoid installing ruby-2.3.1. But we're overwriting this
variable, so its value is not relevant.

I'm reverting the change so the galaxy code remains unchanged, the same
way we don't change the code in a Ruby gem.