Remove explicit libcurl dependencies
The package `libcurl3` is available on Ubuntu 16.04 but is not available
on Ubuntu 18.04. And the package `libcurl4` is available on Ubuntu 18.04
but not on Ubuntu 16.04.

So one way to make this package compatible with both systems is to
install `curl` and `libcurl4-openssl-dev`, which will depend on the
right libcurl package in both distributions.