Add support for Ubuntu 18.04
Note we need to use RVM 1.28.0 in the Travis build to properly unload
RVM before starting. Travis uses RVM 1.29.0 on Ubuntu 18.04, and this
version has a bug which destroys the PATH environment variable when
using `__rvm_unload`.

Ideally we would run the build in a Travis environment which doesn't
install RVM nor Ruby at all. I haven't found a way to do so, not even
using `language: minimal`.