Add experimental Errbit support
The option to install Errbit is still experimental because we're still
not sure whether it's a good idea to run Errbit on the same server as
the application is running, and so it isn't enabled by default.

Note executing `rails runner` in Errbit results in many warnings and
info messages, and so we're using the `mongo` command to get the app
api_key, which we would normally obtain with `App.last.api_key`.

Also note you need to add the `errbit_domain` to your DNS configuration
so the page is correctly accessed, and the installer assumes it will run
on the same machine and use the same SSL certificate as the CONSUL
application.

Finally, make sure you copy the password Errbit will generate during the
installation process for the default Errbit user.