Remove preinstalled versions of PostgreSQL if any
To be sure we are going to run the latest available PostgresSQL version
on each OS.

Travis Ubuntu images have some PostgreSQL pre-installed versions.
As we are installing the latest by ourselves from apt repositories, we
can safely remove pre-installed versions.

Previously to this change, to have multiple versions of Postgres was
causing some flake specs at some Travis jobs because sometimes the job
launched the right PostgresSQL version but others times not.