Explicitly install shared-mime-info
The package is required by the version of mimemagic CONSUL now uses. In
the end it looks like we won't be able to remove the mimemagic
dependency for a while because both paperclip and caxlsx depend on it,
so we need to install this package.

While most GNU/Linux distributions install it by default, some
production servers might not have it installed.

See also commit 00587bc64d.