Serve compressed CSS and JavaScript assets
In commit c01f67ec1 we introduced several nginx templates. Two of them
had this configuration, and one of them didn't.

We chose the one without this configuration for unrelated reasons, and
didn't realize this configuration was missing until now.

Rails already generates a gzipped version of our assets when running
`rake assets:precompile`, so the Nginx configuration is the only code we
need to change to enable this feature.

In CONSUL version 1.3.0, the size of the uncompressed JavaScript file is
1.63MB, while the gzipped file size is 478KB. Similarly, the size of
the uncompressed CSS file is 538KB, while the gzipped file size is 97KB.

So, in all, with this change the size of the homepage decreases from
2.76MB to 1.16MB, meaning a reduction of about 60%.

Since browsers cache CSS and JavaScript assets, this size reduction will
only affect new users, users who have disabled their browser's cache, or
users who haven't visited the page since the last time the assets
changed. Still, new users (particularly the ones with poor internet
connections) will surely enjoy a better experience.