Make sure Nginx has write access to the puma socket
Maybe due to a change in the Ubuntu images provided by Github Actions,
we've started to get a 502 error when checking the page in our
continuous integration service.

Debugging showed the puma socket was created with 775 permissions, as
opposed to the 777 permissions we find when we run the installer on
production machines.

So, as a workaround, we're giving the puma socket the same permissions
it used to have.

I haven't been able to reproduce this issue on production; the added
code is probably redundant in this scenario.