Use systemd to start Errbit
Errbit has updated its version of Puma to Puma 5.x, which removes the
`--daemon` option and requires using systemd or similar.

Our errbit service file is based on the Puma documentation to configure
systemd [1]. We're also adding a dependency on Mongodb and making sure
the process runs as the user with permissions to run Errbit.

Note we're using `Type=simple`, since using `Type=notify` with
`WatchdogSec=10` resulted in the service not returning control to the
shell after being started, which in turn resulted in a timeout error.

Also note we aren't running Errbit as a user service due to problems
with systemd when becoming a different user with Ansible and IMHO adding
the machinectl method just for this task isn't really worth it.

[1] https://github.com/puma/puma/blob/21e9a4a65b3/docs/systemd.md