Fix mini_racer compilation when installing Errbit
Errbit has recently updated its mini_racer dependency to version 0.6.2.
As mentioned in the rubyjs/mini_racer repo issue 243, there seems to
be an issue with rubygems (issue 4488 in rubygems/rubygems) which
results in the libv8-node-16.10.0.0-x86_64-linux-musl.gem being
installed instead of installing the libv8-node-16.10.0.0-x86_64-linux
gem.

We're applying a suggestion given in the comments of the mentioned
mini_racer issue, which makes we'll install the libv8-node gem for the
right platform.