Install PostgreSQL extensions on a shared schema
We've included the Apartment gem in CONSUL, which needs to have a
separate schema with the shared extensions so they're available to every
tenant.

Since the database user CONSUL uses will have no superuser permissions,
we need to add the extensions to the shared schema from the beginning,
right after we create the database.

Note that, once we release a version of CONSUL which includes
multitenancy, we can remove the "Grant usage" section, as it will be
executed as part of the Rails migrations.