Upgrade RVM Ruby galaxy to version 2.2.0
Now we don't have to overwrite the GPG servers anymore, since the ones
we were using are the default ones in the new version.

I've considered using a git submodule instead of downloading the
release, but decided against it because using git submodules would
require an extra step in order to run the installer (running the command
`git submodule update --init`).