Fix Consul 1.5.0 racc installation on Debian CI
It looks like the racc gem is calling the `/usr/bin/mkdir` command. In
some systems, this is a symbolic link to `/bin/mkdir`, while in some
other systems it doesn't exist.

So we're making sure we create the symbolic link if it doesn't exist, as
it seems to be the case in the Docker images for Debian we use in our
CI.