Split Debian tests by Debian version
With this change, we will run Debian tests in separate workflows by
version, so if one job from a Debian version fails, it does not cancel
all the other jobs.

We keep the failing buster tests because the installer works fine on
real servers. It only fails in the CI.