Update tests status badge in README
We renamed the `tests` workflow to `ubuntu` in commit f346954ea, when we
added support for Debian, but forgot to update the badge.

Note that we're only adding the build status on Ubuntu because the
status badge for the Debian build is shown as failing even when tests
are passing. Since we're about to release version 2.1.1, we're leaving
it like that for now, but ideally we'll add the Debian badge in the
future as well.