Only use fnm exec when necessary
Just like we do in the Puma systemd service, we can set `EXECJS_RUNTIME`
to `Disabled`, and so we don't need to include fnm in the commands. We
only need FNM to install Node, install Node packages and precompile
assets, so those are the only places where we're keeping `fnm exec`.
This is similar to what we do when deploying with Capistrano.