Install Ansible with apt in Ubuntu workflows
We were having problems with the python-setup action when trying to use
Ubuntu 24.04 because it uses Python 3.12 instead of Python 3.8 and
because the way to install ansible with pip has changed over the years.

So we're now installing ansible with apt, just like we do in Debian
workflows. This will also install the python3 package if it isn't
already installed.